requirements.txt
# Core dependencies for n-body simulation
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0

# Configuration file support
PyYAML>=6.0

# Optional dependencies for enhanced functionality
# Uncomment if needed:
# h5py>=3.0.0          # For HDF5 data export
# pandas>=1.3.0        # For data analysis
# pillow>=8.0.0        # For animation export